+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 - Added netParams.correctBorder to compensate distance-dependent connectivity border effect
 
+- Added option to run single job on multiple mpi cores via Batch class 
+
 # Version 0.7.8
 
 - Added netParams.rotateCellsRandomly option to rotate cells randomly around y-axis